Refactor: homr_evaluator.py の役割ごとのモジュール分割
- homr_evaluator.py の肥大化を解消するため、内部ロジックを src/homr_eval_scripts/core/ 配下の各モジュール (utils, heuristics, metrics, reporting, predictor) に分割。
- 推論のみを直接実行する際に不要な評価関連のモジュールを読み込まないように依存関係を整理。
- パイプライン (src/pipeline/detection/hybrid.py) 側の import を新しいパスに追従させた。
- Issue #92 の対応
